SpreadsheetEngineContext.sort removed
- Moved all BasicSpreadsheetEngineContext.sort logic into BasicSpreadsheetEngine.sort
